requirements.txt
alembic==1.0.10
Click==7.0
Django==2.2.1
django-admin==1.3.2
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.9.4
Flask==1.0.3
Flask-Cors==3.0.7
flask-heroku==0.1.9
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
psycopg2==2.8.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
Werkzeug==0.15.4
xlwt==1.3.0
